# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Baselix API
  version: 1.0.0
  description: |
    Baselix is a dynamic data-storage API built on an Entity–Attribute–Value (EAV) model.
    You define tables, manage their schemas, and perform full CRUD operations on records —
    all through a simple REST API.

    ## Authentication

    Every API request must include a project API key as a Bearer token:

    ```
    Authorization: Bearer <api_key>
    ```

    API keys are generated per project and can be rotated from the dashboard.

    ## Field Types

    When defining a schema, each field is assigned a type string:

    | Type | Description |
    |------|-------------|
    | `string` | Text |
    | `int` | Integer |
    | `float` | Decimal number |
    | `bool` | Boolean |
    | `time` | ISO 8601 timestamp |
    | `json` | JSON object or array |
    | `uuid` | UUID string |

    Append `_u` to any type (e.g. `string_u`) to enforce a **uniqueness constraint** on that field.
    Only `string`, `int`, `float`, and `uuid` support the `_u` suffix.

    ## Filtering & Sorting

    ### Filter syntax
    ```
    ?filter=field:operator:value,field:operator:value
    ```

    Supported operators: `eq`, `ne`, `gt`, `gte`, `lt`, `lte`, `contains`

    You can filter on user-defined fields as well as the meta fields `id`, `created_at`, and `updated_at`.
    Use `null` as the value for null checks (e.g. `?filter=name:eq:null`).

    ### Sort syntax
    ```
    ?sort=field:direction,field:direction
...

...
servers:
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: integer
          example: 400
        message:
          type: string
          example: "table 'users' already exists"
...
    RecordObject:
      type: object
      description: >
        A record. Includes system-managed metadata (`id`, `created_at`, `updated_at`)
        plus all user-defined field values as top-level keys.
      properties:
        id:
          type: string
          format: uuid
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: true
      example:
        id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        created_at: "2026-04-12T10:00:00Z"
        updated_at: "2026-04-12T10:00:00Z"
        name: "Alice"
        age: 30
        is_active: true
paths:
...
  /tables/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
        description: Table name
        example: users

    get:
      summary: List records
      description: |
        Returns all records from a table, optionally filtered and sorted.

        **Filter syntax:** `?filter=field:operator:value,field:operator:value`

        Supported operators: `eq`, `ne`, `gt`, `gte`, `lt`, `lte`, `contains`

        **Sort syntax:** `?sort=field:direction,field:direction`

        You can filter and sort on user-defined fields as well as `id`, `created_at`, and `updated_at`.
      operationId: getRecords
      tags: [Records]
      parameters:
        - name: filter
          in: query
          required: false
          schema:
            type: string
          description: >
            Comma-separated filter conditions in `field:operator:value` format.
          example: "age:gte:18,name:contains:alice"
        - name: sort
          in: query
          required: false
          schema:
            type: string
          description: >
            Comma-separated sort directives in `field:direction` format.
          example: "created_at:desc,name:asc"
      responses:
        "200":
          description: List of records
          content:
            application/json:
              schema:
                type: object
                properties:
                  records:
                    type: array
                    items:
                      $ref: "#/components/schemas/RecordObject"
              example:
                records:
                  - id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    created_at: "2026-04-12T10:00:00Z"
                    updated_at: "2026-04-12T10:00:00Z"
                    name: "Alice"
                    age: 30
        "400":
          description: Invalid filter or sort format, or filter references a field that does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: 400
                message: 'invalid filter format: invalid filter operator: "contains"'
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Table not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Create record(s)
      description: |
        Creates one or more records. Accepts either a single JSON object or an array of objects.

        **Auto-schema inference:** If the table does not yet exist, it will be created automatically
        using a schema inferred from the supplied data. When this happens, the response includes a
        `message` field and uses HTTP 201.

        Reserved fields (`id`, `created_at`, `updated_at`) must not be included — they are managed
        automatically.
      operationId: createRecords
      tags: [Records]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/RecordObject"
                - type: array
                  items:
                    $ref: "#/components/schemas/RecordObject"
            examples:
              single:
                summary: Single record
                value:
                  name: "Alice"
                  age: 30
                  is_active: true
              multiple:
                summary: Multiple records
                value:
                  - name: "Alice"
                    age: 30
                  - name: "Bob"
                    age: 25
      responses:
        "201":
          description: Record(s) created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Present only when the table was auto-created.
                  records:
                    type: array
                    items:
                      $ref: "#/components/schemas/RecordObject"
              examples:
                existing_table:
                  summary: Inserted into existing table
                  value:
                    records:
                      - id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        created_at: "2026-04-12T10:00:00Z"
                        updated_at: "2026-04-12T10:00:00Z"
                        name: "Alice"
                        age: 30
                new_table:
                  summary: Table auto-created
                  value:
                    message: "Table 'users' created with inferred schema."
                    records:
                      - id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        created_at: "2026-04-12T10:00:00Z"
                        updated_at: "2026-04-12T10:00:00Z"
                        name: "Alice"
                        age: 30
        "400":
          description: Reserved field in payload, type mismatch, or duplicate unique value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                reserved_field:
                  summary: Reserved field
                  value:
                    error: 400
                    message: "cannot use reserved field: id, we manage these fields automatically for you: id, created_at, updated_at"
                duplicate_unique:
                  summary: Duplicate unique field value
                  value:
                    error: 400
                    message: "duplicate value: 'alice@example.com' for unique field: email"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      summary: Update record(s)
      description: |
        Partially updates one or more records. The request body must be an **array** of objects.
        Each object must include an `id` field (UUID) identifying the record to update, plus any
        fields to change.

        Reserved fields `created_at` and `updated_at` cannot be included.
        `updated_at` is automatically refreshed on every successful update.
      operationId: updateRecords
      tags: [Records]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                allOf:
                  - type: object
                    required: [id]
                    properties:
                      id:
                        type: string
                        format: uuid
                  - $ref: "#/components/schemas/RecordObject"
            example:
              - id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                age: 31
                is_active: false
      responses:
        "200":
          description: Updated records
          content:
            application/json:
              schema:
                type: object
                properties:
                  records:
                    type: array
                    items:
                      $ref: "#/components/schemas/RecordObject"
              example:
                records:
                  - id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    created_at: "2026-04-12T09:00:00Z"
                    updated_at: "2026-04-12T11:00:00Z"
                    name: "Alice"
                    age: 31
                    is_active: false
        "400":
          description: Reserved field included, invalid ID format, or duplicate unique value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Table or record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tables/{name}/{id}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
        description: Table name
        example: users
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Record UUID
        example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"

    get:
      summary: Get a record by ID
      description: Returns a single record by its UUID.
      operationId: getRecordByID
      tags: [Records]
      responses:
        "200":
          description: Record found
          content:
            application/json:
              schema:
                type: object
                properties:
                  records:
                    $ref: "#/components/schemas/RecordObject"
              example:
                records:
                  id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  created_at: "2026-04-12T10:00:00Z"
                  updated_at: "2026-04-12T10:00:00Z"
                  name: "Alice"
                  age: 30
        "400":
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: 400
                message: "invalid record ID"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: 404
                message: "record not found"

    delete:
      summary: Delete a record by ID
      description: Permanently deletes a single record and all its stored values.
      operationId: deleteRecordByID
      tags: [Records]
      responses:
        "200":
          description: Record deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "record deleted"
        "400":
          description: Invalid UUID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: 400
                message: "invalid record ID"
        "401":
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: 404
                message: "table 'users' has no record with ID '3fa85f64-5717-4562-b3fc-2c963f66afa6'"